Do not write queue name file stub
The question that we want to find a quick answer to is: Is this really
the reason why we are writing many GBs to disk, and bloat Erlang memory
to 100GB (and beyond) for writing ~200MB worth of message data to disk?

The goal of this commit is to produce a container image that we can test
a specific workload against.

The comment above the commented function has more context.

For the full story, check out https://github.com/rabbitmq/rabbitmq-server/discussions/2785

Everything compiles, but the tests are taking more than 5 minutes to run
on a 20-core Xeon W with 64GB of DDR4 & 3TB/s NVMe drive. I will let
them run, but not let myself get blocked on these completing.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>