First few perf-test runs
It currently performs worse than rabbit_queue_index,
about twice less messages get published and/or consumed.
But that's expected as the changes in rabbit_variable_queue
are still pretty hackish.

It might also make sense to keep the recent N publish in memory
(non-converted to binary as well) so that when reading they are
already there without having to do any file access or conversion.
This would make immediate reads after a publish super fast
without having to keep too much in memory. This cache could
easily be dropped should the read_marker stray off too far
from the write_marker.

The read_ahead feature would do wonders otherwise. But it
currently does not work well with seeks.

Memory when only publishing is around 42k to 52k. When
consuming it grows a lot, but that's because the code
crashes when trying to use --qos and the segments are
larger (65536 entries).