Logging: Standardize global and per-category Cuttlefish variables
With this patch, it is possible to configure categories the same way the
global logging outputs are configured.

For instance:
    log.channel.file = /var/log/rabbitmq/channels.log
    log.channel.file.formatter = json

As an added bonus, it is now possible to set the global log level
without messing with the "default" "category":
    log.level = notice

The compatibility is preserved.

TODO: Extend the Cuttlefish schema to allow to configure the exchange,
journald and syslog outputs for categories as well.