More work to improve the performance of the modern index
When initially publishing and the message was already delivered,
the IsDelivered flag is given immediately rather than in a
separate call. This means we write it to disk when we write
the entry in the index directly.

The read/write were converted to pread/pwrite where applicable.
But the real optimization will come from buffering the writing
of entries, and reading the from this buffer. In that case and
assuming the consumer is fast the only thing we'll ever write
to disk will be ack bytes.